Add macOS dylibs
These dylibs were built from the github source today. The dylibs were renamed and the `lib` prefix was removed. This allows the same binding string to be used for Windows and macOS.

The search paths for the shared libaries needed to be updated as well.

sphinxbase.dylib was modified as follows:
- `install_name_tool -id "@loader_path/sphinxbase.dylib" sphinxbase.dylib`

pocketsphinx.dylib was modified as follows:
- `install_name_tool -id "@loader_path/pocketsphinx.dylib" pocketsphinx.dylib`
- `install_name_tool -change "/usr/local/lib/libsphinxbase.3.dylib" "@loader_path/sphinxbase.dylib" pocketsphinx.dylib`